move organism repository to warehouse package
