+++ src/main/java/org/sbml4j/repository/warehouse/OrganismRepository.java
@@ -11,7 +11,7 @@
  * 
  * For a full list of authors, please refer to the file AUTHORS.
  */
-package org.sbml4j.repository.common;
+package org.sbml4j.repository.warehouse;
 
 import org.sbml4j.model.warehouse.Organism;
 import org.sbml4j.model.warehouse.WarehouseGraphNode;
